+++ docs/source/changelog.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+jsonyx 2.3.0 (unreleased)
+-------------------------
+
+- Added ``colored`` to :class:`jsonyx.Encoder`, :func:`jsonyx.dump`,
+  :func:`jsonyx.dumps` and :func:`jsonyx.write`
+
 jsonyx 2.2.0 (Mar 31, 2025)
 ---------------------------